refactor: refactored folder structure
