+++ src/features/inventory/component/stock-slider/stock-slider.tsx
